refactor: apply code review
